Reordered resources to ensure base layouts are for the smallest supported screens
Both layout/layout-port distinctions were due to lack of horizontal space. Converted to using layout for the base (smaller horizontal width) layouts and layout-w720dp for the larger horizontal width layouts. This ensures that both landscape and portrait 10" layouts are the same.

Note that we need to use layout aliases in this case to ensure that Android 3.1 devices are not adversely affected.