Remove race condition on Event updating, ensure correct update increment
As Event updating only occurred after a database operation, it was possible to hit 'New Game' before it completed, leading to an error as the GoogleApiClient was no longer connected. Also the value that was being used was the number of games played since last sync, rather than the number of games _won_ since last sync.